major changes to improve controls:
add target on the move gadget
change default "attack" command to "settarget", bind a to "canceltarget"
add widget to handle cmd colors directly (note: there is no api for this)

This means the gravity gun keeps its target and fires while you move, making it much easier to control where you push/pull stuff.  The target can be dropped by pressing a (canceltarget) or s (stop). The current target, if there is one, is displayed with a blue line stipple.